made the number of meetings to populate on the left column be calculated dynamically
Starting from the year 2021, some mwb publications contain the presentations and parts
for 2 consecutive months. This is a problem for wrex because the number of meetings to
populate on the first (left) column of the Excel document is hard-coded at 3 (3 on the
left and 8 on the right. Yikes!). To solve this problem, instead of hard-coding the count
in the code, we dynamically calculate it before starting the population. This commit does
not complete the feature but this branch will.